% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_control.R
\name{baseline_control}
\alias{baseline_control}
\title{Control for baseline differences}
\usage{
baseline_control(
  pupil,
  time,
  baseline_period,
  grouping = NULL,
  log = FALSE,
  log_file = NULL
)
}
\arguments{
\item{pupil}{A numeric vector containing pupil data.}

\item{baseline_period}{A numeric vector that specifies the two time periods}

\item{grouping}{A vector containing grouping information}
}
\description{
\code{baseline_control} returns a vector of pupil data
}
\examples{

}
