% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pupils.R
\name{combine_pupils}
\alias{combine_pupils}
\title{Combine the left and right pupil observations}
\usage{
combine_pupils(
  pupil_left,
  pupil_right,
  grouping = NULL,
  log = FALSE,
  log_file = NULL
)
}
\arguments{
\item{pupil_left}{A numeric vector containing pupil data.}

\item{pupil_right}{A numeric vector containing pupil data.}

\item{grouping}{A vector containing grouping information}
}
\description{
\code{combine_pupils} returns the average between the left and right pupil.
}
\examples{

}
