% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log.R
\name{create_log}
\alias{create_log}
\title{Create a log of pupil data transformations}
\usage{
create_log(location = NULL, file_name = NULL, overwrite = FALSE)
}
\arguments{
\item{location}{A character string specifying the location where the log
should be created.}

\item{file_name}{A character string specifying the name of the file. By default
this is set to 'pupil-preprocessing'}

\item{overwrite}{A boolean specifying whether the log file should be
overwritten.}
}
\description{
\code{create_log} creates a file that can be used to log the pupil
preprocessing steps.
}
