% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_gaps.R
\name{describe_gaps}
\alias{describe_gaps}
\title{Calculate summary statistics of gaps in the data}
\usage{
describe_gaps(data, time, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{One or more numeric columns in the data frame that you want to
calculate the gap descriptives of.}

\item{timestamp}{The timestamp column in the data frame.}
}
\description{
\code{describe_gaps} returns a data frame with summary statistics describing
the gaps in the data.
}
\details{
Returns a data frame containing the start and end point, as well as
the duration of each gap.

The data set can be grouped using \strong{dplyr}'s \code{group_by}
so that the gap descriptives will be calculated for each group level.
}
\examples{
library(dplyr)

describe_gaps(gaps, pupil_left)
describe_gaps(gaps, pupil_left, pupil_right)
gaps \%>\%
  group_by(trial) \%>\%
  describe_gaps(pupil_left, pupil_right)

}
