% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_missing.R
\name{describe_missing}
\alias{describe_missing}
\title{Describe missing data statistics}
\usage{
describe_missing(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{A variable number of numeric columns to calculate the missing
descriptives of.}
}
\description{
\code{describe_missing} returns a data frame with summary statistics
describing the missing data of both or either pupil
}
\details{
The data set can be grouped using \strong{dplyr}'s \code{group_by}
so that the missing descriptives will be calculated for each group level.
}
