% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_validity.R
\name{describe_validity}
\alias{describe_validity}
\title{Calculate summary statistics of validity indicators}
\usage{
describe_validity(data, ..., log = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{...}{One or more numeric columns in the data frame that contain data on
the validity of the recordings.}

\item{log}{A boolean indicating whether the result is added to the log.}

\item{message}{A string containing an optional message to add to the log.}
}
\description{
\code{describe_validity} returns a data frame with summary statistics
describing the eye tracker's validity indicators
}
\details{
The data set can be grouped using \strong{dplyr}'s \code{group_by}
so that the gap descriptives will be calculated for each group level.
}
