% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{log_message}
\alias{log_message}
\title{Log message}
\usage{
log_message(message, log_file = NULL)
}
\arguments{
\item{message}{A character string specifying the message.}

\item{log_file}{A character string specifying the path to the log file.}
}
\description{
\code{log_message} adds a message to the pupil preprocessing log.
}
\examples{

}
