% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_removal.R
\name{log_removal}
\alias{log_removal}
\title{Log removal of observations}
\usage{
log_removal(
  data,
  left_eye = NULL,
  right_eye = NULL,
  left_eye_old = NULL,
  right_eye_old = NULL,
  log_file = NULL,
  message = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{left_eye}{A numeric column in 'data' containing the left eye
observations.}

\item{right_eye}{The name of the column in 'data' containing the right eye
observations}

\item{left_eye_old}{An optional column specifying the reference column for
the left eye}

\item{right_eye_old}{An optional column specifying the reference column for
the right eye}

\item{log_file}{A character string specifying the path to the log file.}
}
\description{
\code{log_removal} updates the pupil preprocessing log with the number of
observations that were removed.
}
\examples{

}
