% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_section.R
\name{log_section}
\alias{log_section}
\title{Log section}
\usage{
log_section(section, log_file = NULL)
}
\arguments{
\item{section}{A character string specifying the section name.}

\item{log_file}{A character string specifying the path to the log file.}
}
\description{
\code{log_section} adds a section to the pupil preprocessing log.
}
\examples{

}
