% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dilation_speed_outliers.R
\name{remove_dilation_speed_outliers}
\alias{remove_dilation_speed_outliers}
\title{Remove dilation speed outliers}
\usage{
remove_dilation_speed_outliers(pupil, time, constant = 10)
}
\arguments{
\item{pupil}{A numeric vector of pupil size changes.}

\item{time}{A vector containing the timestamps associated with the
pupil size measurements.}

\item{constant}{An integer that raises or lowers the cutoff threshold.}
}
\description{
\code{remove_dilation_speed_outliers} removes pupil size outliers based on
dilation speed.
}
\details{

}
\examples{
library(dplyr)

blink <- mutate(blink,
    pupil_left_clean = remove_dilation_speed_outliers(pupil_left, time,
      constant = 10),
    pupil_right_clean = remove_dilation_speed_outliers(pupil_right, time,
      constant = 10)
  )

}
