% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothen.R
\name{smoothen}
\alias{smoothen}
\title{Smoothen pupil measurements}
\usage{
smoothen(
  pupil,
  time = NULL,
  method = "low pass",
  sampling_frequency = NULL,
  cutoff_frequency = 10,
  padding = 50,
  span = 0.2
)
}
\arguments{
\item{pupil}{A numeric vector containing pupil data.}

\item{method}{The type of interpolation to be used. Possible values are
'low pass' (default) or 'loess'.}
}
\description{
\code{smoothen} returns a vector of smoothened pupil measurements according
to the specified method.
}
\examples{

}
