% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_line_outliers_to_na.R
\name{trend_line_outliers_to_na}
\alias{trend_line_outliers_to_na}
\title{Trend line outliers to missing.}
\usage{
trend_line_outliers_to_na(
  pupil,
  time = NULL,
  constant = 10,
  method = "low pass",
  sampling_frequency = 60,
  cutoff_frequency = 10,
  padding = 50,
  interpolation = "linear",
  log = FALSE,
  log_file = NULL
)
}
\arguments{
\item{pupil}{A numeric vector of pupil size measurements.}

\item{time}{A vector containing the timestamps associated with the
pupil size measurements.}

\item{constant}{A numeric value specifying the threshold for outlier removal.}

\item{log}{A logical value. Should the action and results be logged?}

\item{log_file}{A character string specifying the path to the log file.}

\item{span}{A numeric value specifying the amount of smoothing.}
}
\description{
\code{trend_line_outliers_to_na} determines outliers based on deviations from
a smooth trend line and sets the outliers to missing (NA).
}
\examples{
library(dplyr)

blink <- mutate(blink,
    pupil_left = trend_line_outliers_to_na(pupil_left, timestamp),
    pupil_right = trend_line_outliers_to_na(pupil_right, timestamp)
  )

}
