% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_log.R
\name{update_log}
\alias{update_log}
\title{Update the pupil preprocessing log}
\usage{
update_log(location = NULL, file_name = NULL, overwrite = FALSE)
}
\arguments{
\item{location}{A character string specifying the location where the log
should be created.}

\item{file_name}{A character string specifying the name of the file. By default
this is set to 'pupil-preprocessing'}

\item{overwrite}{A boolean specifying whether the log file should be
overwritten.}
}
\description{
\code{update_log} updates the pupil preprocessing log.
}
\examples{

# Validity
# Outliers
# Gaps
# Sparsity
# Smoothing

}
