% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{view_log}
\alias{view_log}
\title{View the pupil preprocessing log}
\usage{
view_log(log)
}
\arguments{
\item{log}{A variable containing the pupil preprocessing log, created with
\code{create_log}.}
}
\description{
\code{view_log} prints the pupil preprocessing log to the console.
}
\examples{
See \code{\link[eyepatch]{create_log}} for an example.

}
