% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{write_log}
\alias{write_log}
\title{Write the pupil preprocessing log to a file}
\usage{
write_log(log, file)
}
\arguments{
\item{log}{A variable containing the pupil preprocessing log, created with
\code{create_log}.}

\item{file}{A character string specifying the path and file name.}
}
\description{
\code{write_log} saves the log to a file on your computer.
}
\examples{
See \code{\link[eyepatch]{create_log}} for an example.

}
