% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{create_log}
\alias{create_log}
\title{Create a log of pupil preprocessing steps}
\usage{
create_log()
}
\description{
\code{create_log} creates a variable can be updated to log the pupil
preprocessing steps and how these steps affected the pupil data.
}
\details{
Use the \code{cat} function to print the log to the console in a
human-friendly format.
}
\examples{
# Create the log
log <- create_log()



# Print the log
cat(log)

}
