% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gaps.R
\name{interpolate_gaps}
\alias{interpolate_gaps}
\title{Interpolate missing data between gaps}
\usage{
interpolate_gaps(x, type = "linear", rule = 2, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{The type of interpolation to be used. Possible values are
'linear' (default) or 'spline'.}
}
\description{
\code{interpolate_gaps} returns a vector with the (originally) missing data
points interpolated according to the specified method.
}
\details{
Returns a vector equal in length to \code{x}, with all missing data
points interpolated according to the specified method.

See \code{?approx} and \code{?spline} for more available arguments.

Note that the default value for the rule argument in \code{approx()} has
changed from 1 to 2.
}
\examples{
library(dplyr)
library(ggplot2)

# Linear interpolation
gaps <- mutate(gaps, pupil_left_LI = interpolate_gaps(pupil_left))

ggplot(gaps, aes(x = timestamp)) +
 geom_point(aes(y = pupil_left_LI), color = "red") +
 geom_point(aes(y = pupil_left))

# Spline interpolation
gaps <- mutate(gaps, pupil_left_spline = interpolate_gaps(pupil_left,
  type = "spline"))

ggplot(gaps, aes(x = timestamp)) +
 geom_point(aes(y = pupil_left_spline), color = "red") +
 geom_point(aes(y = pupil_left))

}
