% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_to_na.R
\name{invalid_to_na}
\alias{invalid_to_na}
\title{Set invalid measures to missing}
\usage{
invalid_to_na(pupil, condition, log = FALSE, log_file = NULL)
}
\arguments{
\item{pupil}{A numeric vector containing pupil data.}

\item{condition}{A condition to determine which values should be set to NA.}

\item{log}{A logical value. Should the action and results be logged?}

\item{log_file}{A character string specifying the path to the log file.}
}
\description{
\code{invalid_to_na} sets invalid measures to missing (NA).
}
\examples{

# Set 4 to missing in a vector from 1 to 10
invalid_to_na(1:10, 1:10 == 4)

# Set all pupil sizes with a value of -1 to NA
invalid \%>\%
  mutate(
    pupil_left = invalid_to_na(pupil_left, pupil_left == -1),
    pupil_right = invalid_to_na(pupil_right, pupil_right == -1)
  )

# Or, set pupil sizes to NA based on a validity score

invalid \%>\%
  mutate(
    pupil_left = invalid_to_na(pupil_left, validity_left == 4),
    pupil_right = invalid_to_na(pupil_right, validity_right == 4)
  )


}
