% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_section}
\alias{log_section}
\title{Add a section in the pupil preprocessing log}
\usage{
log_section(log, section)
}
\arguments{
\item{log}{A variable containing the pupil preprocessing log, created with
\code{create_log}.}

\item{section}{A character string specifying the section name.}
}
\description{
\code{log_section} adds a section to the pupil preprocessing log.
}
\details{
\code{log_section} automatically turns the section name into a
level-2 Markdown header and separates it from preceding messages with two
line breaks.
}
\examples{
See \code{\link[eyepatch]{create_log}} for an example.

}
