% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing}
\alias{missing}
\title{Missing timestamp data}
\format{

}
\usage{
missing
}
\description{
Pupil size data from a single trial that contains implicit missing timestamps.
}
\keyword{datasets}
