% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_gaps.R
\name{pad_gaps}
\alias{pad_gaps}
\title{Pad gaps}
\usage{
pad_gaps(
  pupil,
  time,
  gap_minimum = 75,
  padding_before = 50,
  padding_after = 50,
  grouping = NULL,
  log = FALSE,
  log_file = NULL
)
}
\arguments{
\item{pupil}{A numeric vector of pupil size measurements.}

\item{time}{A vector containing the timestamps associated with the
pupil size measurements.}

\item{gap_minimum}{A numeric value describing the minimum gap duration}

\item{padding_before}{A numeric value describing by how much gaps should be
padded before a gap}

\item{grouping}{A vector containing grouping information}

\item{log}{A logical value. Should the action and results be logged?}

\item{log_file}{A character string specifying the path to the log file.}
}
\description{
\code{pad_gaps} removes samples that border certain gaps in the data.
}
\examples{
library(dplyr)

gaps <- mutate(gaps,
  pupil_left_padded = pad_gaps(pupil_left, timestamp, gap_minimum = 10,
    padding_before = 4, padding_after = 4),
  pupil_right_padded = pad_gaps(pupil_right, timestamp, gap_minimum = 10,
    padding_before = 4, padding_after = 4)
  )

ggplot(gaps, aes(x = timestamp)) +
  geom_point(aes(y = pupil_left), color = "red") +
  geom_point(aes(y = pupil_left_padded))

}
