% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_to_na.R
\name{sparse_to_na}
\alias{sparse_to_na}
\title{Sparsity filter}
\usage{
sparse_to_na(
  pupil,
  time,
  gap_criterion = 40,
  cluster_criterion = 50,
  log = FALSE,
  log_file = NULL
)
}
\arguments{
\item{pupil}{A numeric vector of pupil size measurements.}

\item{time}{A vector containing the timestamps associated with the
pupil size measurements.}

\item{gap_criterion}{A numeric value specifying the gap duration preceding
potential sparse observations.}

\item{cluster_criterion}{A numeric value specifying the section duration of
sparse value that should be removed.}

\item{log}{A logical value. Should the action and results be logged?}

\item{log_file}{A character string specifying the path to the log file.}
}
\description{
\code{sparse_to_na} removes sparse samples in the data.
}
\examples{
library(dplyr)

gaps <- mutate(gaps,
  pupil_left_clean = sparse_to_na(pupil_left, timestamp,
    gap_criterion = 9, cluster_criterion = 11),
  pupil_right_clean = sparse_to_na(pupil_right, timestamp,
    gap_criterion = 9, cluster_criterion = 11),
  )

ggplot(gaps, aes(x = timestamp)) +
  geom_point(aes(y = pupil_left), color = "red") +
  geom_point(aes(y = pupil_right), color = "red") +
  geom_point(aes(y = pupil_left_clean)) +
  geom_point(aes(y = pupil_right_clean))

}
